pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "senatus"
version = "0.0.1"
description = "Senatus: a trading and market analisys lib"
authors = [
  { name = "Marco Penaforte", email = "penafortemarco@gmail.com" }
]
readme = "README.md"
license = { file = "LICENSE.txt" }
requires-python = ">=3.10"
keywords = ["trading", "market-analysis", "quant", "arcana", "roman"]

classifiers = [
  "Programming Language :: Python :: 3.10",
  "License :: OSI Approved :: MIT License",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Operating System :: OS Independent",
]

dependencies = [
  "numpy>=1.23",
  "pandas>=1.5",
  "plotly>=0.0.1",
  "binance-connector>=3.12.0",
  "binance-futures-connector>=3.12.0",
]

[tool.setuptools]
packages = ["senatus"]

